+++ src/main/java/io/supertokens/oauth/OAuth.java
@@ -16,11 +16,15 @@
 
 package io.supertokens.oauth;
 
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import io.supertokens.Main;
 import io.supertokens.config.Config;
 import io.supertokens.httpRequest.HttpRequest;
 import io.supertokens.httpRequest.HttpResponseException;
 import io.supertokens.oauth.exceptions.OAuthAuthException;
+import io.supertokens.oauth.exceptions.OAuthClientRegisterException;
 import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.StorageUtils;
 import io.supertokens.pluginInterface.exceptions.InvalidConfigException;
@@ -28,11 +32,14 @@ import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
 import io.supertokens.pluginInterface.oauth.OAuthStorage;
+import io.supertokens.pluginInterface.oauth.exceptions.OAuth2ClientAlreadyExistsForAppException;
 import io.supertokens.utils.Utils;
 
 import java.io.IOException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
 import java.util.*;
 
 public class OAuth {
@@ -43,7 +50,7 @@ public class OAuth {
     private static final String ERROR_DESCRIPTION_LITERAL = "error_description=";
 
     private static final String HYDRA_AUTH_ENDPOINT = "/oauth2/auth";
-
+    private static final String HYDRA_CLIENTS_ENDPOINT = "/admin/clients";
 
     public static OAuthAuthResponse getAuthorizationUrl(Main main, AppIdentifier appIdentifier, Storage storage, String clientId,
                                                         String redirectURI, String responseType, String scope, String state)
@@ -90,6 +97,74 @@ public class OAuth {
         }
 
         return new OAuthAuthResponse(redirectTo, cookies);
+    }
+
+    //This more or less acts as a pass-through for the sdks, apart from camelCase <-> snake_case key transformation and setting a few default values
+    public static JsonObject registerOAuthClient(Main main, AppIdentifier appIdentifier, Storage storage, JsonObject paramsFromSdk)
+            throws TenantOrAppNotFoundException, InvalidConfigException, IOException, OAuthClientRegisterException,
+            NoSuchAlgorithmException, StorageQueryException {
+
+        OAuthStorage oauthStorage = StorageUtils.getOAuthStorage(storage);
+        String adminOAuthProviderServiceUrl = Config.getConfig(appIdentifier.getAsPublicTenantIdentifier(), main).getOAuthProviderAdminServiceUrl();
+
+        byte[] idBaseBytes = new byte[48];
+
+        while(true){
+            new SecureRandom().nextBytes(idBaseBytes);
+            String clientId = "supertokens_" + Utils.hashSHA256Base64UrlSafe(idBaseBytes);
+            try {
+                if(oauthStorage.isClientIdAlreadyExists(clientId)){
+                    continue; // restart
+                }
+
+                JsonObject hydraRequestBody = constructHydraRequestParamsForRegisterClientPOST(paramsFromSdk, clientId);
+                JsonObject hydraResponse = HttpRequest.sendJsonPOSTRequest(main, "", adminOAuthProviderServiceUrl + HYDRA_CLIENTS_ENDPOINT, hydraRequestBody, 10000, 10000, null);
+
+                oauthStorage.addClientForApp(appIdentifier, clientId);
+
+                return formatResponseForSDK(hydraResponse); //sdk expects everything from hydra in camelCase
+            } catch (HttpResponseException e) {
+                String errorMessage = e.rawMessage;
+                JsonObject errorResponse = (JsonObject) new JsonParser().parse(errorMessage);
+                String error = errorResponse.get("error").getAsString();
+                String errorDescription = errorResponse.get("error_description").getAsString();
+                throw new OAuthClientRegisterException(error, errorDescription);
+            } catch (OAuth2ClientAlreadyExistsForAppException e) {
+                //in theory, this is unreachable. We are registering new clients here, so this should not happen.
+                throw new RuntimeException(e);
+            }
+        }
+
+    }
+
+    private static JsonObject constructHydraRequestParamsForRegisterClientPOST(JsonObject paramsFromSdk, String generatedClientId){
+        JsonObject requestBody = new JsonObject();
+
+        //translating camelCase keys to snakeCase keys
+        for (Map.Entry<String, JsonElement> jsonEntry : paramsFromSdk.entrySet()){
+            requestBody.add(Utils.camelCaseToSnakeCase(jsonEntry.getKey()), jsonEntry.getValue());
+        }
+
+        //add client_id
+        requestBody.addProperty("client_id", generatedClientId);
+
+        //setting other non-changing defaults
+        requestBody.addProperty("access_token_strategy", "jwt");
+        requestBody.addProperty("skip_consent", true);
+        requestBody.addProperty("subject_type", "public");
+
+        return requestBody;
+    }
+
+    private static JsonObject formatResponseForSDK(JsonObject response) {
+        JsonObject formattedResponse = new JsonObject();
+
+        //translating snake_case keys to camelCase keys
+        for (Map.Entry<String, JsonElement> jsonEntry : response.entrySet()){
+            formattedResponse.add(Utils.snakeCaseToCamelCase(jsonEntry.getKey()), jsonEntry.getValue());
+        }
+
+        return formattedResponse;
     }
 
     private static Map<String, String> constructHydraRequestParamsForAuthorizationGETAPICall(String clientId,